backend/requirements.txt
anyio==3.7.0
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
fastapi==0.97.0
h11==0.14.0
httptools==0.5.0
idna==3.4
mysql-connector-python==8.0.33
passlib==1.7.4
protobuf==3.20.3
pydantic==1.10.9
PyJWT==2.7.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3